Fix several issues with the GOp mission UI
Fixes #16 and #17

The interaction between the base `UIMission` and mission-specific
subtypes changed in WotC and the LW version needed some updates. This
change goes through the `UIMission_GOps` class from WOTC and updates the
LW version to be roughly equivalent.

In particular the class needs an empty `OnButtonSizeRealized` function
to avoid the base version moving the confirm button around (#16), and
the screen building needs to invoke the base class `BuildScreen` to
properly set up the chosen and sitrep UI elements (among others) so
they will be hidden when appropriate.

It's probably better to split this file up into separate versions for
different mission types, but I didn't do that here. Non-GOp missions
may still have bugs because of this. Dark Event display may need
updating too.