Fix Jailbreak and main retaliation missions
This mainly involved fixing the Kismet so that civilians/rebels switched
sides correctly. It also makes sure that any liaison on the mission is
actually controllable by the player. The problem had the same cause as
the team swapping for rebels: the liaison wasn't being added to XCOM's
initiative group.