Fix the aliens' aim
The LW2 config had some odd values for settings that affected soldier
and alien aim. The idea was to work around problems that the "aim
adjusting" code in vanilla caused for the graze band. But we're not
using the graze band currently and the special values were completely
messing aim values up.

This change simply restores those config settings to sane values for
vanilla behaviour.