Fix rebel raid missions (mini-retals)
The rebels weren't being added to XCom's initiative group at the start
of the mission, so they weren't controllable by the player. There is now
a dedicated utility function that does this and all the bits of the mod
that need to add rebels etc. to XCom at the start of the mission now use
it.

Note that Rendezvous missions add the rebels direct to the XCom squad,
so they don't need to use this new function.