Restore modified Golden Path missions
The LW version of the final "Fortress" mission was crashing the game on
startup. It appears that WOTC split the mission in two and probably
changed a bunch of other stuff, such as including the Chosen. The old
replacement clearly wasn't a match.

Hopefully the other Golden Path missions are OK, but I haven't tested
them yet.