Hide reinforcements flares/force show reinforcements alert
Implements both #34 and #36. The community highlander has been updated with
some events that allow us to disable the reinforcements flares and also
force the display of the reinforcements alert.

The reinforcements alert changes mean that we can now show the green/yellow/
red alerts based on how long before the reinforcements actually arrive.