Fix black market to always use supplies (#78)
The initial configuration of the black market to use supplies didn't require
any changes the highlander, but the end of month restock did. This change
relies on a new `BlackMarketGoodsReset` event in the community highlander to
customise the inventory.

With the event in place and by attaching a listener to `OnGeoscapeEntry`, I
was able to remove all the hard-coded black market update code in the DLCInfo
class.