Disable buggy plot (#58)
`Plot_SLM_Overpass_LgObj_10_LW` relies on some PCPs that were in the original
game but were removed from WOTC. It looks like the overpass PCPs were
refactored into smaller ones.

Anyway, it's too big a job to fix the plot on probably not worth it, so
disabling it for now.