Slight balance changes and renamed folder
Renamed the folder WotCWeaponsAndArmor and LW_WeaponsAndArmor for consi
sistency between other mod names that start with "LW_".

Renamed class X2Item_Weapons to X2Item_FactionWeapons.

Renamed class X2Item_Armors to X2Item_FactionArmors.

Deleted the Schematics class X2Item_Schematics as the behavior to
remove schematics was just a simple matter of removing the lines of
the schematics in the LW_Overhaul.ini file.

Gave credit to LW2 developer in class X2Ability_FactionWeaponAbilities
for use of the code for SMG abilities for added mobility to SMGs and
applied it to Bullpups.

Added faction weapons to ItemTable in LW_Overhaul.ini to set up costs
and slots and to set other properties like making them finite.