Restore selected GTS upgrades (#123)
Several of the GTS upgrades had been moved to resistance order card. This meant
that some crucial upgrades were out of the player's control.

This change restores the following upgrades and removes them as cards:

 * Vulture
 * Vengeance
 * Integrated warfare
 * Wet work
 * Lightning strike

Note that the cards haven't exactly been removed. It's just that their strength
is so high that they will never be available to the player.