Improve display of custom soldier statuses
This takes advantage of some changes in the community highlander that
allow us more control over how statuses and their durations are
displayed. For example, Haven Adviser now doesn't show any duration and
if a soldier is too low rank for haven duty, that is shown in red.

I have also added a duration for the "Infiltrating" status, which
displays how long is left to reach 100% infiltration. It does not factor
in mission expiry, but still provides a useful idea of how long you
might need to wait before a soldier becomes available again, or whether
even to abort a current infiltration.