Fix primary weapon selection in soldier loadout
Switching weapons on soldiers was failing to work in that the game always
switched to the weapon you previously selected before a new selection. So
the selected weapon always lagged the player's selection by one.

This was due to a misunderstanding in how the
`CanAddItemToInventory_CH_Improved()` hook worked. The implementation of
that method should not return true if the slot already contains a weapon.
The code no longer overrides the default vanilla logic except for the case
of disallowing certain heavy weapons for certain soldier classes.