Unable to infiltrate alien fdacility fix
This push is to fix the issue of not being able to infiltrate alien
facilities (Issue #139). You can now infiltrate a facility and complete
it successfully. It will remove the amount of doom the facility has but
will not play the animation and display the notification of having done
so. This isa WIP commit and will need to figure out the hook to call to
play the animation.