Unable to infiltrate alien facility fix (#139) (#163)
This push is to fix the issue of not being able to infiltrate alien
facilities (Issue #139). You can now infiltrate a facility and complete
it successfully. It will remove the amount of doom the facility has but
will not play the animation and display the notification of having done
so (work that still needs to be done).