Shadow Chamber Issues Fix (#140)
The Shadow Chamber had 2 issues. The first was that it displayed
information for missions that were not Golden Path missions or Alien
Facility missions which it was not supposed to do. The second issue
was that it set alert level to all missions to 1 which caused baseline
enemy activity to be wrong (too low) for a lot of the regions come late
game.