Second part of the Shadow Chamber fix.
The initial shadow chamber fix only fixed the issues that it was
causing for missions that were not infiltrating. For missions that were
infiltrating the changes that were made in class UIMission_LWCustomMis
sion needed to be made in class UIMission_LWLaunchDelayedMission.uc as
well because the issue cropped back up when missions were being infiltr
ated and that class handles missions that are being infiltrated.