Cleaned up code in BindLibraryItem() function
Cleaned up some of the code in the below classes:
    UIMission_LWCustomMission
    UIMission_LWLaunchDelayedMission.uc

I cleaned up the code for the default case of the swtich statement that
had all of the code from the parent class pasted in minus the code that
displayed shadow chamber issue. I put the call to parent class verion
of BindLibraryItem() for each and then just called the Hide() function
on the ShadowChamber panel for each class.