Modify the Smash N Grab mission (#53)
The mission is now an evac mission with a fixed evac zone, so no corpse
recovery. In addition, when the aliens first see XCOM, they call for
reinforcements, for which a timer pops up.

The basic idea currently is that you sneak in, mark as many crates as
possible, then undertake a fighting retreat.