Remove FacelessCivilians from Rendezvous schedules (#63)
Faceless civilians were spawning as part of the mission schedule and this was
causing issues with lack of corpses and activations of the actual Rebel
Faceless.

I have kept FacelessCivilian in the spawn distribution list in case it's
needed for adding those units later, but I have set the spawn weight to zero
so that though don't spawn as normal pods. This allows the normal pods to
spawn as expected.