Fix the dark events UI (#102)
WOTC changed the dark events UI so that the LW2 customisation was no longer
working (done with `UIScreenListener_AdventOperations`). This commit creates
a subclass of `UIAdventOperations` (called `UIAdventOperations_LW`) and uses
a mod class override to replace the original with the custom version.

The custom version manages the UI list of the dark events and ensures that
only one card (the selected one) is ever displayed.