Haven Adviser Fix (#159)
This fix is for fixing the issue of haven advisers not getting any
abilities at all (not from perks, weapons, items, etc..). Looks like
Firaxis added a BeginTacticalPlay() function to some of the XCGS
classes that has to be called to properly load in the abilities.
This fifx just added the function calls in the proper places.