Haven Adviser Fix (#159) (#192)
This fixes the issue of haven advisers not getting any abilities
at all (not from perks, weapons, items, etc..).

Looks like Firaxis added a `BeginTacticalPlay()` function to
some of the XCGS classes that has to be called to properly
load in the abilities. This fix just adds the function calls where
we add haven advisers to missions —
`LWTacticalMissionUnitSpawner.LoadLiaisonFromOutpost()`.