Add event listener for avatar project progress
Add an event listener that utilizes an event hook that listens for an
event fired whenever doom is updated. The event listener adds up all
of the doom for any facilities that have not been revealed and adds
that to all of the doom that has currently been gathered so that the
amount of doom is updated properly. The issue before this was that
alien facilities that were not revealed had an issue when updating doom
in that when they added a point of doom that point would be removed
then added which would make it look like it did not move up at all but
then when the facility was revealed it would add up all of the doom
that the facility had actually added.