Fix SPARK healing while infiltrating (#232)
LW2 was using the soldier status `eStatus_Healing` to determine whether
a SPARK going on a mission needed to be removed from the repair slot.
Unfortunately, that status was not reliably set.

This fixes #232 by overriding the delegate functions used to add SPARKs
to the repair slot and remove them. The new functions just additionally
set the soldier status. It also fixes a bug in LW2 where it was
overriding the `eStatus_Healing` status when it shouldn't.