Add event listener for avatar project progress (#231)
Add an event listener to handle whenever doom is updated. The event listener
adds up all of the doom for any facilities that have not been revealed and
adds that to all of the doom that has currently been gathered. This ensures
that the doom counter properly reflects the current Avatar progress.

The issue was that unrevealed alien facilities had an issue when updating doom.
Whenever they added a point of doom that point would be removed, then added.
Then when the facility was revealed, the game would add up all of the doom
that the facility had actually accumulated and display that in the doom meter.
In other words, a potentially big blob of doom would suddenly appear to the
player in the Avatar progress UI.

Fixes #100.