Fix for burnout ability #233
The ability burnout was not updating the tile that the soldier was
standing in to give it the smoke cover bonus but was updating the tiles
adjacent to the unit. For some reeason the ability tempaltes multi
target effect for smoke only updated the adjacent tiles. This was fixed
by adding a single target effect that applied to the tile that the
soldier was on.