Fix haven advisers disappearing after Defend (#222)
The code that was processing units after retaliation missions was checking
each unit's `bRemovedFromPlay` flag. That's OK for rebels and MECs, but all
XCOM squad members' flags are already cleared by that point by the
`XComGameStateContext_StrategyGameRule.SquadTacticalToStrategyTransfer()`
function.

The haven adviser is now treated separately, so if they aren't dead or
captured, then the assumption is that they got out of the mission OK.

Fixes #222.