Enable various AI abilities (#129)
LW2 cleared out the `EquivalentAbilities` config array which was
critical to the AI being able to select certain abilities. It broke Holy
Warrior for example because `HolyWarriorM1`, `HolyWarriorM2`, and
`HolyWarriorM3` were no longer aliased to the `HolyWarrior` ability that
the AI behaviour trees were checking for.

It seems that LW2 did this just to add `LightEmUp` as an alias to
`StandardShot`, so that is now more explicit in the config.

Fixes #129.