Fix remaining yellow alert reflex action issues (#283)
Most of the work for this issue was done in the highlander, but this
commit integrates with those changes by updating the event that LWOTC
listens for (from 'ScamperBegin' to 'ProcessReflexMove') and configuring
the reflex action point types added by LWOTC so that the game considers
them for reflex actions (rather than ignoring them).

Fixes #283.