Integrate Perfect Information for LW2
This adds important information to the shot HUD (when players are
selecting targets to shoot) that breaks the hit chance down into graze
and crit chances.

The behaviour has been changed slightly in that the To Hit chance now
includes the graze chance on top of the crit chance. Previously it just
included the crit chance. So now the displayed hit chance is the chance
to graze, hit normally, or crit.

I've also completely redone the way the flyover information is handled.
It now only works for StandardAim-based abilities, but that's the most
important use case anyway. And additionally, we no longer override the
base game `X2Action_ApplyWeaponDamageToUnit`, which means that the
Extended Information! mod can work with LWOTC if a player wants to use
that. Still need to set up a some sort of config option to allow players
to disable the Perfect Information HUD and flyover though.