Refactor game state effect components (#52,#64)
I have removed all the code associated with game state effect
components, since they are apparently liable to crash the game. This
substantial change meant updating a host of effects, and in some cases
fixing them.

With the help of the LW2 Classes and Perks mod, I was also able to
identify the issue with Trojan: the TrojanVirus effect was using the
'PlayerTurnBegun' event, but that no longer fires at the appropriate
time. The effect now uses the 'UnitGroupTurnBegun' event instead.

Lastly, I updated Whirlwind to use the same mechanic as Templar's
Momentum.

Implements #52, fixes #64.