Fix psi ops not getting promotions
They were missing out on mission XP because there was some code that
triggered one of the XP-related events, but the name of that event has
since changed. So the listener that handled mission XP wasn't being
called for psi ops.

The code now goes through the standard `XCGS_Unit.GetTotalNumKills()`,
which triggers the correct event.