Fix `SetSquadCrew()` (#213)
Apparently `SetSquadCrew()` on `XCGS_LWPersistentSquad` is used from
multiple places for different purposes, which means you have to be
careful what you do in it. In this case, the fix to display the correct
squad in the dropship/other mission loading screens broke some squad
management.

The solution was to only apply the fix if the `bOnMissionSoldiers`
argument is `true`.

This fixes #213, #302, and #314.