Restore Red Fog stats on evac (#319)
The red fog effect extends `X2Effect_ModifyStats`, but it wasn't calling
the super implementation of the `OnEffectRemoved()` listener function. I
now call that super implementation because it cleans up the stat
modifiers that were otherwise lingering on soldiers after they returned
from missions.

Note that this didn't affect wounded soldiers at the end of a successful
sweep objective mission because that situation is handled by a different
listener function that we don't override in `X2Effect_RedFog_LW`.

Fixes #319.