Changed SMG Weapon Category (#317)
SMGs were getting the damage bonus from the assault rifle weapon
breakthrough. This was because their weapon category was 'rifle' which
is the same category as the assualt rifle. Changed the weapon category
to 'smg' and added the needed lines in XComClassData.ini to allow the
classes that could equipem them to continue to equip them.