Fix code that adds Chosen to missions (#374)
The logic for adding Chosen to missions was being overridden by base
game code. This change fixes that and, because the Chosen are only added
to missions on launch now, only the first mission after activation
guarantees their appearance (as well as Golden Path missions).

I have also adjusted the appearance rate to a base 20%, which is
adjusted by the infiltration amount. So infiltration above 100% reduces
the chance of the Chosen appearing, while underinfiltrating makes them
more likely to appear. And if you underinfiltrate enough, you guarantee
that a Chosen will appear on the mission (unless the mission type is
explicitly excluded).

Fixes #374.