Refactor the pod manager for WOTC (#408)
WOTC introduced initiative groups and the 'UnitGroupTurnBegun' event.
These broke the pod manager because it did all its work in a
'PlayerTurnBegun' listener, before the units were initialised with their
action points. The lack of action points was messing up the calculations
for where yellow alert pods should move to, so they ended up staying
put.

This refactor puts the pod job initialisation and execution code in a
'UnitGroupTurnBegun' listener where it needs to be done now. I have also
moved the listeners into an `X2EventListener` class so they don't
persist in the game state (there's no reason for them to).

Fixes #408.