Fix reinforcements alert (#401)
Apparently searching in a game state for a game state object does not
automatically search in the history. So when the
'OverrideReinforcementsAlert' event changed to pass in the associated
game state, I shouldn't have used solely that to find the
`XCGS_AIReinforcementSpawner`. This change looks in the history, but now
includes the associated game state in the search (as I understand it).

Fixes #401.