Block some techs from inspiration (#91)
The rendering techs, datapad and data cache research, and facility leads
can no longer be inspired. This was done by making it possible to
research them, but only if you have 1000 Gatekeeper corpses. No specific
reason for that particular requirement other than it's very, very
unlikely to happen in a campaign.

Resolves #91.