Faction Unit Rebalance
This commit introduces the rebalancing of the faction hero units. It
aims to normalize the classes a bit more to make all 3 as equally
viable as possible. The short version of this update is that the
templar got some new abilities but no too many as they did not need
much in the way of abilities, just some to help round them out.
Reaper's stealth was toned down quite a bit with their detection radius
being about a tile lower than smg but they start with ghostwalker. Made
remote start a charged base ability with only 1 charge and now they
have reasonable stealth and are still a good pick but are not the
unstopable rampaging killing machines they were before. The skirmishers
recieved quite a bit of love and seem to be in a very good spot now.
You can have 3 different types of skirmishers all 3 of which mainly
support the squad with completing objevctives. They can deal ok damage
but are mainly focused around utlity now.