Fix rocket targeting (#441)
This appears to have been an old bug from the original LW2 in which the
rocket targeting checked both the unit tile and peek tile for line of
sight on the target. But in the second instance, it actually checked the
line of sight between the unit tile and the peek tile!

I have changed this to check LoS from the peek tile to the target tile
as it really should be.

Fixes #441.