Fix activation of the Chosen
I made a mistake when switching to a Second Wave option for this and
left the check in the template create code. Of course, the second wave
option is tied to the campaign, so this is completely the wrong point at
which to test for it.

I have simply added a completion function for the Assault Network Tower
objective that always runs, but it performs the check on the Second Wave
option and activates the Chosen if it's enabled.