Update missions to capture soldiers (#238)
Several of the mission umap files weren't capturing XCOM soldiers at the
end of a successful mission. That's probably because the vanilla
missions they're based on are sweep, in which case capture makes no
sense.

Anyway, I updated all the necessary missions, so your soldiers can now
be captured at the end of successful missions.

Fixes #238.