Update to pull request
Resolved some of the issues raised for the pull request and also made
some changes to the skirimisher Battlemaster ability. Removed the
abilities tradecraft, damn good ground and suppression from it. Now
it has the damn good ground effect built into it and set its covertness
in XComLW_InfiltrationSettings.ini to match that of Tradecraft. Also
fixed their ability tree in ClassData.ini so that heavy ordinance is
not on their 2nd tier random tree twice.