Make first liberation mission obvious
Hiding which mission is the start of the Liberation chain is just
unfriendly to newcomers, while experienced players have a pretty good
idea which mission is Lib 1 (since the only other mission that looks
like it is detect a landed UFO). This makes it clearer that the Lib 1
mission actually starts you down the path of liberating the region.