Adjust Skirmisher abilities
Forward Operator has been temporarily removed because it can be highly
abused in its current implementation. It needs a fix that requires
testing, which will do later.

Reckoning doesn't really fit with the theme of the Skirmisher, who's
more about utility and mobility rather than melee damage. Plus it wasn't
getting the Fleche damage bonus. We can reconsider this one later.

Finally, the Run And Gun has been replaced with the LW variant so that
it can't be used after Hit And Run procs.

Fixes #504.