Fix science score when scientists are busy (#511)
Scientists assigned to the Infirmary or assigned as haven advisers were
not being removed from the calculated science score because the base
game doesn't check for scientists in staff slots.

The fix for the Infirmary requires a highlander change (see the issue
X2CommunityCore/X2WOTCCommunityHighlander#626). With it, the new
`X2EventListener_Headquarters` has a listener that adjusts the science
score based on any scientists in staff slots.

The haven adviser fix was simple: set the status of scientists and
engineers to `eStatus_CovertAction`, which WOTC does take account of.
Note that soldier haven advisers also have this status.

Fixes #511.