Fix and modify Secure UFO mission (#595)
I have updated the Secure UFO mission so that the umap is based on the
new WOTC version of the mission. I have also rejigged the reinforcements
so that it works with the continuous reinforcements subsystem, which can
now be explicitly triggered, for example when a mission timer runs out.

Note that the mission switches from a sweep objective to an evac
objective if the beacon timer runs out before it's disabled. This means
that the resources and corpses are not collected.

Fixes #595.