Fix some dark events
An error in the INIs meant that the unit types that are supposed to be
affected by some dark events, like Lightning Reflexes, weren't. A
mismatch between the property names in the code and the ones in the INI
files led to the unit type arrays being empty.

I've fixed the property names in XComGameCore.ini.