Apply crit chance debuff for Silent Killer
If a Reaper has Silent Killer and is concealed, they receive a
significant debuff to their crit chance. This is to offset the fact they
can one shot tough enemies without losing concealment because of their
high crit damage and Silent Killer. Once out of concealment, the debuff
is lost.