Make correct Chosen appear in each Stronghold (#439)
I was clearing the Chosen tags from all missions, because parts of the
base game were adding the tags when I didn't want them to. However, that
meant that the Chosen tag was also being removed from the final mission
and the Chosen Stronghold missions.

This change simply skips modifying the Chosen tags if the mission is one
of those just mentioned.

Fixes #439.