Fix Trojan Virus removing action points (#605)
The effect was removing the action points too soon because the listener
that did it was activating on *all* the `UnitGroupTurnBegun` events, not
just the one for the corresponding hacked unit.

This simply makes sure that the current unit group turn matches the unit
with the Trojan Virus before doing its work and removing the effect.