Fix duration of Mind Merge
Mind Merge was being removed at the beginning of the alien turn, which
made it pretty useless. I think this was because the Lost now go before
the aliens and that broke the implementation. I've made the
implementation more "normal" now so that it lasts for a full turn (Lost,
Aliens, Neutral + Resistance).