Try to fix crash with Rocket Launchers mod
It seems that LWOTC crashes with Iridar's Rocket Launchers mod, possibly
because `LWTemplateMods` attempts to use a custom damage effect as the
source for a copy constructor that expects a simple
`X2Effect_ApplyWeaponDamage` instance.

Note that I don't know if this actually fixes the problem because I
couldn't reproduce it in the first place. However, this replicates some
behaviour around grenade-like _abilities_.