Fix Sting grenades appearing in inventory (#442)
This actually applies to any temporary item, such as dense smoke
grenades. The problem was that units that were removed from play before
the end of the mission weren't having their temporary items removed.

I've fixed this by adding an `OnEffectRemoved()` listener to
`X2Effect_TemporaryItem`. It seems both it and the `OnTacticalGameEnd()`
listeners are needed.

Fixes #442.