Increase rank of first faction soldiers (#700)
This change overrides the `GiveRewardFn` function for the Find Faction
and Find Farthest Faction reward templates so that the corresponding
covert actions reward higher rank factions soldiers than before. The
rank of each faction soldier is equal to one less than the rank
requirement for the covert action. So the second faction soldier you get
will be CPL and the third will be SGT (the first one you get on
Gatecrasher will still be Squaddie).

Resolves #700.