Make Revival Protocol Fixes a required mod
This is basically to make sure that Revival Protocol and Restoration
clear the stunned effect from units. Otherwise Revival Protocol bugs out
on units that are both disoriented and stunned, removing the disorient
and granting two action points, but leaving the unit stunned. So you can
move the unit, but it can't do anything else.

Plus, the description for Revival Protocol says it clears stunned.