Clear stunned, etc. when mind control lost (#627)
This is another attempt to fix the mind control + stun issue. This
change adds a 'MindControlCleanse' ability that removes disoriented,
stunned, and other effects from a unit when it is no longer mind
controlled. This mirrors the behaviour of clearing those effects when
the unit is first mind controlled.

Requires a new highlander event: 'MindControlLost'.

Fixes #627.