Recover haven adviser from non-evac retals (#768)
The code that recovers XCOM bodies and equipment at the end of missions
wasn't triggering for units that were added to the mission as if they
had spawned from the Avenger. This mostly affected haven advisers, who
could die at the end of missions if they were bleeding out.

This change duplicates the necessary code from X2TacticalGameRuleset,
putting it in the custom tactical mission cleanup code of LWOTC,
applyinh the same processing for haven advisers as for the normal XCOM
squad soldiers.

Fixes #768.