Ensure Lost spawn on Abandoned City maps (#765)
Several of the mission types disabled the Lost reinforcement system if
there were no more Lost on the map. Since LWOTC doesn't put any Lost on
the map in the first place, this meant that the Lost RNFs were disabled
from the start.

This fixes the Kismet in Destroy Relay, Hack Workstation and Recover
Item missions to *not* disable the Lost.

Fixes #765.