Rework implementation of tactical dark events
The previous implementation used hack reward templates to apply dark
event abilities like Shredder and Lethal, but that meant they were
applied *after* `FinalizeUnitAbilitiesForInit()`, meaning they couldn't
be patched. In addition, they weren't applied to units dropped into the
map via the `DropUnit` console command.

The new implementation works in a similar way to the vanilla tactical
dark events like Lightning Reflexes:

 1. A new ability called 'DarkEventAbility_<RealAbilityName>' is created,
    e.g. 'DarkEventAbility_Lethal'
 2. The new ability adds its corresponding ability to the
    `AdditionalAbilities` array
 3. The ability includes conditions based on whether the dark event
    tactical gameplay tag is active, whether the unit meets certain
    criteria, and whether a successful chance-based roll is made
 4. This ability is added to all the enemy character templates that it is
    configured for

The dark event ability itself is hidden, but the ability it adds is
visible on the units for which it's active.