Fix Close Combat Specialist on Vanguard (#699)
The CCS ability needs to be associated with a weapon or at least a
weapon slot, but that information can't be included when giving an
ability to a character template, which is how we add abilities to enemy
units.

This change patches a configurable list of abilities in
`FinalizeUnitAbilitiesForInit()` such that those abilities are tied to
the unit's primary weapon. This doesn't just fix CCS, but also Center
Mass, Lethal, etc.

Fixes #699.