Add commands for respeccing soldiers
I have added `RespecSelectedSoldier`, which respecs whatever soldier is
selected in the Armory UI, and `RespecAllSoldiers`. These commands are
designed for when the game updates the ability trees for soldiers. The
soldiers go back to Squaddie, but keep their previous XP. This means
that can be promoted back to their original rank using the updated
ability tree.