Add new Distraction ability for LWOTC
When a Reaper has this passive ability, their Claymores and homing mines
apply an additional disoriented effect to any units caught in the
explosion.

The Claymore and homing-mine explosions are handled differently. The
latter is simply modified to apply disoriented if the shooter has the
Distraction ability.

The Claymore needed more work. It now has a short chain of listeners
that are designed to react to when the Claymore explodes and apply the
disoriented effect around where the Claymore is at that point in time.
There is also a merge visualisation function that ensures the
disoriented flyover message appears *after* the explosion.