Refactor Fleche effect and add cap
The bonus damage per tile is no longer a config var in the effect
itself, but has to be initialised by the ability that applies the Fleche
effect. This enables different abilities to have different
damager-per-tile values.

The Fleche bonus damage is now also capped to prevent some exploits
(with Icarus Armor for example) that previously resulted in huge damage
numbers.