Fix Serial to not crash in debug (#282)
The code that removes effects at the end of turns has an assertion that
checks for `EffectName` on the effect that's being removed. The special
effect that LWOTC adds to do the aim and crit reduction didn't have
`EffectName` set, so that was triggering the assertion, and hence
crashing the game when the mod was built with the debug flag.

This change simply gives the effect a name.

Fixes #282.