Fix configuration for Caster Gauntlets.
The extra damage INI configuration was lacking `+` signs at the start of
the lines, despite configuring entries of an array. This was preventing
Caster Gauntlets from doing damage with Volt.

In addition, this fixes the missing right bracer on the Templar pawn
with conventional Caster Gauntlets and it hides the bonus aim from the
armory UI, to make it consistent with other melee weapons like the Shard
Gauntlets and swords.

Finally, I've added the conventional caster gauntlets to the
`UnlimitedItemsAdded` config, added an `ItemTable` entry for them, and
fixed up some template settings.