Add Project Miranda sit rep
This new sit rep applies an aim and mobility malus ("Lethargic") to all
organic units on the map.

This change also removes the `VALID_SIT_REPS` config: we should be able
to rely solely on the `SitReps` array of the base game. This has the
advantage that players can immediately start using sit reps from other
mods without having to mess with the config.

If the player does want some control over what sit reps can be attached
to missions, then they can add entries to the `SIT_REP_EXCLUSIONS`
config array, which is a list of names of sit reps that should never be
encountered.