Make Meditation charge based, but rechargeable
Meditation is now charge based and starts with 0 charges. However, the
Templar gains a charge with every kill (doesn't have to be with Rend).
The number of charges can never go above one, but that charge can be
refreshed with kills any number of times in a mission.

Note that this makes use of XMB.