Play nicely with RPGO's Promotion Screen
This commit takes advantage of a recent community highlander change that
allows us to override the promotion screen via an event rather than an
MCO. This helps with RPGO because it still uses an MCO for now and that
was conflicting with LWOTC's one.

The listener that overrides the promotion screen has a dedicated check
for if RPGO is installed. If it is, LWOTC doesn't try to override the
promotion screen, since RPGO's one is operational.