Fix mission difficulty evaluation code (#140)
This is a replacement for the original #140 fix, which simply avoided
the problem by not showing the Shadow Chamber panel. But in the process
we were skipping the `UIMission.UpdateMissionSchedules()` code, which
was being used by the Alien Rulers DLC to put the Rulers in the Shadow
Chamber information.

The underlying issue here was that WOTC modified UIMission so that it
ultimately called `X2MissionSourceTemplate.GetMissionDifficultyFn` to
evaluate the alert level for the mission. But LW2 effectively had a
dummy implementation that ended up setting the alert level to 1,
regardless of the base region strength and infiltration level.

That mission difficulty function has been reworked to properly evaluate
the current alert level for the mission. This also required a change to
the template so that the alert level cap that WOTC has is ignored. We
definitely don't want an alert level cap of 5!

Fixes #140.