Fix Chosen showing after being disabled mid campaign
The Second Wave option Enable Chosen was ensuring that LWOTC never added
Chosen tactical gameplay tags to missions itself, but vanilla was still
doing it. This change ensures that the Chosen tactical tags are *always*
removed from the mission, and only added back if the Enable Chosen
option is enabled.