Fix Skirmishers and Templars throwing evac flare (#433)
In many instances (don't know the exact parameters required to
reproduce) Skirmishers and Templars weren't getting the correct
targeting visuals when attempting to throw an evac flare. This change
introduces a hack to fix the weapon entity's pawn, which is `none` for
some reason.

Note that this same hack is used by the base grenade targeting to fix an
issue in the tutorial, or so the comments suggest.

Fixes #433.