Fix camera panning on promotion in After Action screen
I was missing an important hook in the highlander that allows mods to
control where the camera is located for the promotion screen when it's
opened on the After Action screen. This change adds the necessary
listener, ensuring that the soldier pawn stays on the right-hand side
rather than in the centre.