Add a covert action for getting corpses
This new covert action (Steal Corpses) mitigates some of the RNG of
getting corpses for research and building items. It uses a bucket of
loot tables from which the reward is selected, with corpses only being
available if the current Force Level means that the corresponding enemy
can be on missions.