Rework Will loss and recovery (#859)
In mission, Will is no longer lost on seeing enemy pods. One point of
Will is now lost every 4 turns. This makes things less random and a bit
easier to understand for players. Note that the Will loss no longer
depends on the soldiers' current Will.

The bigger changes are around Will recovery:

 * Will no longer recovers while units are infiltrating, training or on
   covert actions

 * Will recovery is a bit slower, since Will is lost more slowly than
   before

Fatigue should now be something that builds up slowly, but requires
some real rest time for soldiers to recover it. Also note that because
the Will recovery rate for a soldier is proportional the their max
Will, high-Will soldiers recover the Will lost during a mission quicker
than low-Will soldiers. This means that they will be combat ready for
more of the campaign, which should lead to faster promotions for them.