Revert damage falloff behaviour
The fix for damage falloff actually meant that the centre tile of a
grenade could roll falloff damage. This is unexpected behaviour for the
player from LW2 and is it a pretty hard nerf. So I'm changing the
distance ratio thresholds so that the fixed logic results in the same
behaviour in LW2.

We will ideally fix the damage falloff logic so that it takes into
account tiles, probably by resolving all damage points to the centres of
the tiles that they are in, and then performing the calculations on
those tile centres.