Make failed covert actions display as failed
The UI was not giving any indication that covert actions were failing,
so I have overridden the `UICovertActionReport` class with one that adds
the localised text "FAILED" in a couple of locations. It's not perfect,
but once players know what to look for, they'll be able to easily see
whether a covert action failed or not.