Add a stat-boosting covert action
The new Intense Training covert action allows one soldier to get a stat
boost for the cost of 5 ability points. The covert actions screen has
been updated so that it displays the resource cost of the covert action.

These covert actions can spawn once per faction as normal, but they can
spawn for multiple factions in the same "month".